+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.5.0
 
 * Add support for defining enums.
+* Fix keyword ordering for `const factory` constructors.
 
 ## 3.4.1